+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/main/scala/com/azure/cosmos/spark/ThroughputControlHelper.scala
@@ -4,7 +4,7 @@
 package com.azure.cosmos.spark
 
 import com.azure.cosmos.implementation.ImplementationBridgeHelpers
-import com.azure.cosmos.models.{CosmosBulkExecutionOptions, CosmosChangeFeedRequestOptions, CosmosItemRequestOptions, CosmosQueryRequestOptions, PriorityLevel}
+import com.azure.cosmos.models.{CosmosBulkExecutionOptions, CosmosChangeFeedRequestOptions, CosmosItemRequestOptions, CosmosQueryRequestOptionsBase, PriorityLevel}
 import com.azure.cosmos.spark.diagnostics.BasicLoggingTrait
 import com.azure.cosmos.{CosmosAsyncContainer, ThroughputControlGroupConfigBuilder}
 import org.apache.spark.broadcast.Broadcast
@@ -67,7 +67,7 @@ private object ThroughputControlHelper extends BasicLoggingTrait {
     }
 
     def populateThroughputControlGroupName(
-                                              queryRequestOptions: CosmosQueryRequestOptions,
+                                              queryRequestOptions: CosmosQueryRequestOptionsBase[_],
                                               throughputControlConfigOpt: Option[CosmosThroughputControlConfig]
                                           ): Unit = {
         if (throughputControlConfigOpt.isDefined) {