+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/rx/OrderbyDocumentQueryTest.java
@@ -196,15 +196,20 @@ public class OrderbyDocumentQueryTest extends TestSuiteBase {
         String query = String.format("SELECT value r.propInt FROM r where r.propInt != null ORDER BY r.propInt %s", sortOrder);
         CosmosQueryRequestOptions options = new CosmosQueryRequestOptions();
         ImplementationBridgeHelpers
-            .CosmosQueryRequestOptionsHelper
-            .getCosmosQueryRequestOptionsAccessor()
+            .CosmosQueryRequestOptionsBaseHelper
+            .getCosmosQueryRequestOptionsBaseAccessor()
             // Custom Factory Method will always get the ObjectNode - so if VALUE function is used
             // the value needs to be extracted manually. This is intentional right now
             // to allow late-binding the decision whether we really want to surface JsonNode or ObjectNode to
             // customers if we ever make the custom factory method public
             // For now in Spark don't need to worry about extracting values - we would need a wrapper to
             // allow inferring schema anyway.
-            .setItemFactoryMethod(options, (node) -> node.get("_value").intValue());
+            .setItemFactoryMethod(
+                ImplementationBridgeHelpers
+                    .CosmosQueryRequestOptionsHelper
+                    .getCosmosQueryRequestOptionsAccessor()
+                    .getImpl(options),
+                (node) -> node.get("_value").intValue());
 
         int pageSize = 3;
         CosmosPagedFlux<Integer> queryObservable = createdCollection.queryItems(query, options,